requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
requests
fastapi
uvicorn
pydantic
